[1/21] Move code from app to controllers
CatalogController
