+++ src/azul/service/catalog_controller.py
@@ -1,8 +1,15 @@
+from inspect import (
+    signature,
+)
 from typing import (
     Annotated,
+    Any,
 )
 
 import attr
+from more_itertools import (
+    one,
+)
 
 from azul import (
     CatalogName,
@@ -10,6 +17,8 @@ from azul import (
     config,
 )
 from azul.openapi import (
+    format_description as fd,
+    responses,
     schema,
 )
 from azul.plugins import (
@@ -26,6 +35,43 @@ from azul.types import (
 
 
 class CatalogController(ServiceController):
+
+    def handlers(self) -> dict[str, Any]:
+        @self.app.route(
+            '/index/catalogs',
+            methods=['GET'],
+            cors=True,
+            spec={
+                'summary': 'List all available catalogs.',
+                'tags': ['Index'],
+                'responses': {
+                    '200': {
+                        'description': fd('''
+                            The name of the default catalog and a list of all available
+                            catalogs. For each catalog, the response includes the name
+                            of the atlas the catalog belongs to, a flag indicating
+                            whether the catalog is for internal use only as well as the
+                            names and types of plugins currently active for the catalog.
+                            For some plugins, the response includes additional
+                            configuration properties, such as the sources used by the
+                            repository plugin to populate the catalog or the set of
+                            available [indices][1].
+
+                            [1]: #operations-Index-get_index__entity_type_
+                        '''),
+                        **responses.json_content(
+                            # The custom return type annotation is an experiment. Please
+                            # don't adopt this just yet elsewhere in the program.
+                            one(signature(self.list_catalogs).return_annotation.__metadata__)
+                        )
+                    }
+                }
+            }
+        )
+        def list_catalogs():
+            return self.list_catalogs()
+
+        return locals()
 
     # The custom return type annotation is an experiment. Please don't adopt
     # this just yet elsewhere in the program.
